Fixed swin_transformer segmentation fault.

+++ mmcls/models/utils/augment/cutmix.py
@@ -123,7 +123,7 @@ class BatchCutMixLayer(BaseCutMixLayer):
         super(BatchCutMixLayer, self).__init__(*args, **kwargs)
 
     def cutmix(self, img, gt_label):
-        one_hot_gt_label = F.one_hot(gt_label, num_classes=self.num_classes)
+        one_hot_gt_label = F.one_hot(gt_label, num_classes=self.num_classes).float()
         lam = np.random.beta(self.alpha, self.alpha)
         batch_size = img.size(0)
         index = torch.randperm(batch_size)
